feat(sequelize): create separate sequelize file
